requirements.txt
tensorflow
numpy~=1.20
Pillow